requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
docutils==0.21.2
fastapi==0.115.12
fpdf==1.7.2
frozendict==2.4.6
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.24
langchain-core==0.3.55
langchain-ollama==0.3.2
langchain-text-splitters==0.3.8
langgraph==0.3.30
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.31
lxml==5.3.2
markdown2==2.5.3
MarkupSafe==3.0.2
multidict==6.4.3
multitasking==0.0.11
nicegui==2.15.0
numpy==2.2.4
ollama==0.4.8
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
peewee==3.17.9
pillow==11.2.1
platformdirs==4.3.7
propcache==0.3.1
pscript==0.7.7
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-multipart==0.0.20
python-pptx==1.0.2
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==75.8.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
starlette==0.46.2
tenacity==9.1.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.1
uvloop==0.21.0
vbuild==0.8.2
watchfiles==1.0.5
websockets==15.0.1
wheel==0.45.1
wsproto==1.2.0
XlsxWriter==3.2.3
xxhash==3.5.0
yarl==1.19.0
yfinance==0.2.55
zstandard==0.23.0
newspaper3kaiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cssselect==1.3.0
dash==3.0.4
dash-bootstrap-components==2.0.2
docutils==0.21.2
fastapi==0.115.12
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
Flask==3.0.3
fpdf==1.7.2
frozendict==2.4.6
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.24
langchain-core==0.3.55
langchain-ollama==0.3.2
langchain-text-splitters==0.3.8
langgraph==0.3.30
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.31
lxml==5.3.2
lxml_html_clean==0.4.2
markdown2==2.5.3
MarkupSafe==3.0.2
multidict==6.4.3
multitasking==0.0.11
narwhals==1.36.0
nest-asyncio==1.6.0
newspaper3k==0.2.8
nicegui==2.15.0
nltk==3.9.1
numpy==2.2.4
ollama==0.4.8
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
peewee==3.17.9
pillow==11.2.1
platformdirs==4.3.7
plotly==6.0.1
propcache==0.3.1
pscript==0.7.7
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-multipart==0.0.20
python-pptx==1.0.2
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
retrying==1.3.4
setuptools==75.8.0
sgmllib3k==1.0.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
starlette==0.46.2
tenacity==9.1.2
tinysegmenter==0.3
tldextract==5.3.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.1
uvloop==0.21.0
vbuild==0.8.2
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.0.6
wheel==0.45.1
wsproto==1.2.0
XlsxWriter==3.2.3
xxhash==3.5.0
yarl==1.19.0
yfinance==0.2.55
zipp==3.21.0
zstandard==0.23.0
